Have ChangeLog direct developers to the Git log.
For open source Git projects, it's common to rely on the git log rather
than a change log file in the repo.

One advantage is that this information is thus accessible to every method
of viewing the commit (including the output of "git log" itself, and
browsing commits with the GitHub web interface).

Another advantage is that you have a log message for every commit, even if
commits are more fine-grained than tags.

The message is currently in "ChangeLog" because that's where CESM
developers are likely to look (since this is the change log's name in other
CESM externals).