Add remaining unit_testing CMake modules.
CESM_utils.cmake and Compilers.cmake were originally left out because
they relate to the operation of run_tests.py, rather than being easily
reusable in other projects like the other modules here.

However, it's much more convenient for the moment to keep the whole
batch together, and it may be that these modules can be made more
generic in the near future.