Use documentation from unit_testing as headers.
This commit moves documentation from the unit_testing external to the
README and to "headers" consisting of comments at the top of each module.
These include the intended "API" for each module.

This documentation is currently put in comments to make it easier for some
of these files to be copied to other projects. The headers also contain
copyright notices, for much the same reason.