Compatibility fix for Compilers change.
Compilers.cmake is now included at the top of CESM_Macros.cmake for the
unit tests, instead of including CESM_Macros inside Compilers.
CESM_utils needed to be adjusted to handle this, and we can no longer
add the NAG -kind=byte change in Compilers due to the risk of
duplicating this flag in the CESM_Macros commands.