Adding support for Cetus
Adding support for the Blue gene parallel job launcher runjob
that is required to run mpi tests on Cetus. By default the
runjob script is used as the launcher when creating the mpi
test.

However the user can override this by setting the
BGQ_RUNJOB environment variable to point to a launcher that
can be used to launch jobs/tests on Cetus. The user needs
to set the BGQ_RUNJOB environment variable before running
cmake (i.e., Setting the environment variable before running
ctest won't work)