Updating the mpiserial check for mpi-serial2.0
TryMPISERIAL.f90 is supposed to fail to compile with mpi-serial
to let the PIO configure (with CMake) know that it should set its
build for mpi-serial. The previous version compiled successfully
with mpi-serial2.0 . We now use an MPI type not defined in the
mpi serial library instead of equivalence of two MPI types in
the mpiserial library to make sure that the code fails to compile
with the mpiserial library.

Thanks Rob Jacob@ANL for the fix.