Fix unit test build to work with mpi-based pFUnit on yellowstone.
This required adding a number of new options to run_tests.py to set some
machine-specific stuff. This also required ensuring that all tests use
create_pFUnit_test, which now adds an mpirun prefix to the command used to run
tests.

Note that the new option that allows setting PAPI_LIB via cmake-args was not
directly related to switching to mpi-based pFUnit, but rather was needed because
-lpapi has been added to the link line for yellowstone-intel.